% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_in}
\alias{insert_in}
\alias{insert_prod}
\alias{insert_end}
\title{Insert text by shortcuts.}
\usage{
insert_in()

insert_prod()

insert_end()
}
\value{
The text
}
\description{
You can edit the shortcuts to insert texts.
}
\author{
Chai
}
