% Generated by roxygen2 (4.0.1): do not edit by hand
\name{shortcut}
\alias{shortcut}
\title{Create Shortcuts}
\usage{
shortcut(url = read_clip(), file = "", open = TRUE)
}
\arguments{
\item{url}{A url to redirect to.}

\item{file}{The name of the output file.  Defaults to the console.
open A name to use for the file (.html extension).}
}
\description{
Creates an HTML file that redirects to a specified link.
}
\examples{
shortcut("http://www.talkstats.com/")
\dontrun{
shortcut("http://trinker.github.io/card/contact")
}
}

