% Generated by roxygen2 (4.0.1): do not edit by hand
\name{shortcut}
\alias{shortcut}
\title{Create Shortcuts}
\usage{
shortcut(file = "", url = read_clip(), open = TRUE, check.ext = TRUE)
}
\arguments{
\item{file}{The name of the output file.  Defaults to the console.
open A name to use for the file (.html extension).}

\item{url}{A url to redirect to.}

\item{open}{logical.  If \code{TRUE} \code{shortcut} attempts to open the
file in a default browser.}

\item{check.ext}{logical.  If \code{TRUE} check for and add an HTML file
extension to file.}
}
\description{
Creates an HTML file that redirects to a specified link.
}
\examples{
shortcut("http://www.talkstats.com/")
\dontrun{
shortcut("http://trinker.github.io/card/contact")
}
}

